+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Add UDP Support to LoadBalancer service
 * fix(instances): ensure internal IP is set when there is only public one
 
 ### Improvements